feat(#graph) : add callback topics
