+++ app-stream-api/src/main/java/com/dingtalk/open/app/api/graph/GraphAPIMessage.java
@@ -1,4 +1,4 @@
-package com.dingtalk.open.app.api.models.graph;
+package com.dingtalk.open.app.api.graph;
 
 import java.util.Map;
 
@@ -8,13 +8,17 @@ import java.util.Map;
  */
 class GraphAPIMessage {
     private Map<String, String> headers;
-
+    private String body;
     public Map<String, String> getHeaders() {
         return headers;
     }
-
     public void setHeaders(Map<String, String> headers) {
         this.headers = headers;
     }
-
+    public String getBody() {
+        return body;
+    }
+    public void setBody(String body) {
+        this.body = body;
+    }
 }
