+++ app-stream-api/src/main/java/com/dingtalk/open/app/api/graph/GraphAPIRequest.java
@@ -1,14 +1,12 @@
-package com.dingtalk.open.app.api.models.graph;
+package com.dingtalk.open.app.api.graph;
 
 /**
  * @author feiyin
  * @date 2023/9/22
  */
-public class GraphAPIRequest extends GraphAPIMessage{
+public class GraphAPIRequest extends GraphAPIMessage {
 
     private RequestLine requestLine;
-
-    private String body;
 
     public RequestLine getRequestLine() {
         return requestLine;
@@ -16,13 +14,5 @@ public class GraphAPIRequest extends GraphAPIMessage{
 
     public void setRequestLine(RequestLine requestLine) {
         this.requestLine = requestLine;
-    }
-
-    public String getBody() {
-        return body;
-    }
-
-    public void setBody(String body) {
-        this.body = body;
     }
 }
