+++ app-stream-api/src/main/java/com/dingtalk/open/app/api/graph/MediaType.java
@@ -1,4 +1,4 @@
-package com.dingtalk.open.app.api.models.graph;
+package com.dingtalk.open.app.api.graph;
 
 /**
  * @author feiyin
