+++ app-stream-api/src/main/java/com/dingtalk/open/app/api/open/http/HttpConstants.java
@@ -5,13 +5,7 @@ package com.dingtalk.open.app.api.open.http;
  * @date 2023/3/1
  */
 public class HttpConstants {
-
     public static final String METHOD_POST = "POST";
-
-    public static final int CONNECTION_TIME_OUT = 3000;
-
-    public static final int READ_TIME_OUT = 3000;
-
     public static final int STATUS_OK = 200;
     public static final String HEADER_CONTENT_TYPE = "Content-Type";
     public static final String HEADER_ACCEPT = "Accept";
