+++ app-stream-api/src/main/java/com/dingtalk/open/app/api/util/GraphUtils.java
@@ -1,10 +1,10 @@
 package com.dingtalk.open.app.api.util;
 
 import com.alibaba.fastjson.JSON;
-import com.dingtalk.open.app.api.models.graph.GraphAPIResponse;
-import com.dingtalk.open.app.api.models.graph.GraphHeaders;
-import com.dingtalk.open.app.api.models.graph.MediaType;
-import com.dingtalk.open.app.api.models.graph.StatusLine;
+import com.dingtalk.open.app.api.graph.GraphAPIResponse;
+import com.dingtalk.open.app.api.graph.GraphHeaders;
+import com.dingtalk.open.app.api.graph.MediaType;
+import com.dingtalk.open.app.api.graph.StatusLine;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -22,8 +22,7 @@ public class GraphUtils {
     }
 
     public static GraphAPIResponse failed(StatusLine statusLine) {
-        GraphAPIResponse response = baseResponse(statusLine);
-        return response;
+        return baseResponse(statusLine);
     }
 
     private static GraphAPIResponse baseResponse(StatusLine statusLine) {
@@ -33,12 +32,5 @@ public class GraphUtils {
         response.setHeaders(headers);
         response.setStatusLine(statusLine);
         return response;
-    }
-
-    private static GraphAPIResponse baseResponse(int code, String reasonPhrase) {
-        final StatusLine statusLine = new StatusLine();
-        statusLine.setCode(code);
-        statusLine.setReasonPhrase(reasonPhrase);
-        return baseResponse(statusLine);
     }
 }
